rename qt_common to common_qt
