+++ electrum/gui/common_qt/plugins.py
